# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "家庭故事豆 API",
    "description": "家庭故事豆 MVP 的 API 契约与 Mock 联调服务。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/recordings": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "List Recordings",
        "operationId": "list_recordings_v1_recordings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_RecordingData__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "recordings"
        ],
        "summary": "Create Recording",
        "operationId": "create_recording_v1_recordings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CreateRecordingData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/podcasts": {
      "post": {
        "tags": [
          "podcasts"
        ],
        "summary": "Create Podcast",
        "operationId": "create_podcast_v1_podcasts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePodcastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/podcasts/{podcast_id}": {
      "get": {
        "tags": [
          "podcasts"
        ],
        "summary": "Get Podcast",
        "operationId": "get_podcast_v1_podcasts__podcast_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "podcast_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Podcast Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "podcasts"
        ],
        "summary": "Remix Podcast",
        "operationId": "remix_podcast_v1_podcasts__podcast_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "podcast_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Podcast Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemixPodcastRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/recordings/{recording_id}/podcast-render-jobs": {
      "post": {
        "tags": [
          "podcast-render"
        ],
        "summary": "Create Podcast Render Job",
        "operationId": "create_podcast_render_job_v1_recordings__recording_id__podcast_render_jobs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastRenderResultData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/podcast-render-jobs/{job_id}": {
      "get": {
        "tags": [
          "podcast-render"
        ],
        "summary": "Get Podcast Render Job",
        "operationId": "get_podcast_render_job_v1_podcast_render_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastRenderResultData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recordings/{recording_id}/podcast-render-job": {
      "get": {
        "tags": [
          "podcast-render"
        ],
        "summary": "Get Recording Podcast Render Job",
        "operationId": "get_recording_podcast_render_job_v1_recordings__recording_id__podcast_render_job_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastRenderResultData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/podcast-render-jobs/{job_id}/retry": {
      "post": {
        "tags": [
          "podcast-render"
        ],
        "summary": "Retry Podcast Render Job",
        "operationId": "retry_podcast_render_job_v1_podcast_render_jobs__job_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastRenderResultData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/podcast-render-jobs/{job_id}/playback-url": {
      "post": {
        "tags": [
          "podcast-render"
        ],
        "summary": "Create Render Playback Url",
        "operationId": "create_render_playback_url_v1_podcast_render_jobs__job_id__playback_url_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastRenderPlaybackData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "timestamp"
        ],
        "title": "ApiMeta"
      },
...
      "ApiResponse_CreateRecordingData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/CreateRecordingData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[CreateRecordingData]"
      },
...
      "ApiResponse_PodcastData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/PodcastData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[PodcastData]"
      },
...
      "ApiResponse_PodcastRenderPlaybackData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/PodcastRenderPlaybackData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[PodcastRenderPlaybackData]"
      },
      "ApiResponse_PodcastRenderResultData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/PodcastRenderResultData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[PodcastRenderResultData]"
      },
...
      "ApiResponse_list_RecordingData__": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/RecordingData"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[RecordingData]]"
      },
...
      "CreatePodcastRequest": {
        "properties": {
          "moment_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Moment Ids"
          }
        },
        "type": "object",
        "required": [
          "moment_ids"
        ],
        "title": "CreatePodcastRequest"
      },
...
      "CreateRecordingData": {
        "properties": {
          "recording": {
            "$ref": "#/components/schemas/RecordingData"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadTargetData"
          }
        },
        "type": "object",
        "required": [
          "recording",
          "upload"
        ],
        "title": "CreateRecordingData"
      },
      "CreateRecordingRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "original_file_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Original File Name"
          },
          "declared_media_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Declared Media Type"
          },
          "declared_size": {
            "type": "integer",
            "maximum": 524288000.0,
            "exclusiveMinimum": 0.0,
            "title": "Declared Size"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "audio_upload",
              "video_upload",
              "mobile_recording",
              "recording_bean"
            ],
            "title": "Source Type",
            "default": "audio_upload"
          }
        },
        "type": "object",
        "required": [
          "title",
          "original_file_name",
          "declared_size",
          "sha256"
        ],
        "title": "CreateRecordingRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStage": {
        "type": "string",
        "enum": [
          "CREATED",
          "UPLOADING",
          "UPLOADED",
          "PREPROCESSING",
          "TRANSCRIBING",
          "ANALYZING",
          "READY_FOR_SELECTION",
          "GENERATING",
          "COMPLETED",
          "FAILED",
          "CANCELLED",
          "DELETED"
        ],
        "title": "JobStage"
      },
...
      "PodcastData": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "pipeline_version": {
            "type": "string",
            "title": "Pipeline Version"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "moments": {
            "items": {
              "$ref": "#/components/schemas/PodcastMomentData"
            },
            "type": "array",
            "title": "Moments"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/PodcastSegmentData"
            },
            "type": "array",
            "title": "Segments"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recording_id",
          "title",
          "status",
          "version",
          "pipeline_version",
          "metadata",
          "moments",
          "segments",
          "created_at",
          "updated_at"
        ],
        "title": "PodcastData"
      },
...
      "PodcastMomentData": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "position": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Position"
          },
          "start_ms": {
            "type": "integer",
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "position",
          "start_ms",
          "end_ms"
        ],
        "title": "PodcastMomentData"
      },
...
      "PodcastRenderPlaybackData": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          },
          "expires_at": {
            "type": "integer",
            "title": "Expires At"
          },
          "render_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Render Metadata"
          }
        },
        "type": "object",
        "required": [
          "url",
          "download_url",
          "expires_at"
        ],
        "title": "PodcastRenderPlaybackData"
      },
      "PodcastRenderResultData": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/PodcastRenderJobData"
          },
          "podcast_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Podcast Version Id"
          },
          "object_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Key"
          },
          "render_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Render Metadata"
          }
        },
        "type": "object",
        "required": [
          "job",
          "podcast_version_id"
        ],
        "title": "PodcastRenderResultData"
      },
...
      "RecordingData": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "family_id": {
            "type": "string",
            "format": "uuid",
            "title": "Family Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "original_file_name": {
            "type": "string",
            "title": "Original File Name"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          },
          "file_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "audio_upload",
              "video_upload",
              "mobile_recording",
              "recording_bean"
            ],
            "title": "Source Type",
            "default": "audio_upload"
          },
          "status": {
            "$ref": "#/components/schemas/JobStage"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "delete_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delete At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "family_id",
          "title",
          "original_file_name",
          "media_type",
          "file_size",
          "duration_ms",
          "sha256",
          "status",
          "delete_at",
          "created_at",
          "updated_at"
        ],
        "title": "RecordingData"
      },
      "RemixPodcastRequest": {
        "properties": {
          "moment_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Moment Ids"
          },
          "intro": {
            "type": "string",
            "maxLength": 600,
            "minLength": 1,
            "title": "Intro"
          },
          "outro": {
            "type": "string",
            "maxLength": 600,
            "minLength": 1,
            "title": "Outro"
          }
        },
        "type": "object",
        "required": [
          "moment_ids",
          "intro",
          "outro"
        ],
        "title": "RemixPodcastRequest"
      },
...
      "UploadTargetData": {
        "properties": {
          "method": {
            "type": "string",
            "const": "PUT",
            "title": "Method",
            "default": "PUT"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_at": {
            "type": "integer",
            "title": "Expires At"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          }
        },
        "type": "object",
        "required": [
          "url",
          "expires_at",
          "headers"
        ],
        "title": "UploadTargetData"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}